src/auth-service/requirements.txt
flask==3.0.3
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pyjwt==2.9.0
bcrypt==4.2.0

